Make it a lot more modular
