+++ src/js/SubCategory.jsx
@@ -1,0 +1,34 @@
+import React, { useState } from 'react';
+import {
+    categoryIDToValue,
+    categoryIDTosubCategoriesIDs,
+    subCategoryIDToValue,
+} from "./Config";
+
+export function SubCategory({setChosenSubCategory, setChosenContentTitle, chosenCategoryID, chosenSubCategoryID}) {
+    // If the person clicks subcategory, ChosenContentTitle should be back to null too
+    const handleSubCategoryClick = (subCategoryID) => {
+        setChosenSubCategory(subCategoryID);
+        setChosenContentTitle(null);
+    };
+
+    return (
+        <div className='subcategory-rectangle'>
+        {Object.keys(categoryIDToValue).map((categoryID) => {
+            if (chosenCategoryID === categoryID) {
+                const subCategoriesIDGivenACategory = categoryIDTosubCategoriesIDs[categoryID];
+                return ( subCategoriesIDGivenACategory.map((subCategoryID) => {
+                    const subCategoryTitle = subCategoryIDToValue[subCategoryID];
+                    return (
+                        <div className='subcategory-title'
+                            isselected={ chosenSubCategoryID === subCategoryID ?  'true' :  'false' }
+                            key={ subCategoryTitle }
+                            onClick={() => handleSubCategoryClick(subCategoryID)}> {subCategoryTitle}
+                        </div>
+                    );
+                }));
+            }
+        })}
+        </div>
+    );
+}
